pyproject.toml
[project]
name = "dailyai"
version = "2.0.0"
description = "AI News Intelligence Platform — LangGraph powered, NiceGUI frontend"
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Shivang Sinha"},
]

dependencies = [
    # UI Framework
    "nicegui>=2.0",

    # Web
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",

    # LangChain / LangGraph
    "langgraph>=0.2",
    "langchain>=0.3",
    "langchain-google-genai>=2.0",
    "langchain-groq>=0.2",
    "langchain-community>=0.3",
    "langchain-openai>=0.3",

    # Data
    "feedparser>=6.0.11",
    "pydantic>=2.9.0",
    "aiosqlite>=0.20",

    # Infrastructure
    "python-dotenv>=1.0.1",
    "apscheduler>=3.10.4",
    "resend>=2.0.0",
]

[project.scripts]
dailyai = "dailyai.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dailyai"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["__pycache__"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
no_implicit_optional = true
check_untyped_defs = true
ignore_missing_imports = true
exclude = ["__pycache__"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
